backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
jellyfish==1.1.0
lxml==5.3.0
python-dateutil==2.9.0
httpx==0.27.2
bcrypt==4.2.1
dnspython==2.7.0
apscheduler==3.10.4
anthropic>=0.40.0
email-validator>=2.0.0
